+++ Monday.Client/Mutations/CreateItem.cs
@@ -13,7 +13,7 @@
         /// <summary>
         ///     The board's unique identifier.
         /// </summary>
-        public int BoardId { get; set; }
+        public long BoardId { get; set; }
 
         /// <summary>
         ///     The group's unique identifier.
