+++ Monday.Client/Mutations/UpdateColumn.cs
@@ -8,12 +8,12 @@
         /// <summary>
         ///     The board's unique identifier.
         /// </summary>
-        public int BoardId { get; set; }
+        public long BoardId { get; set; }
 
         /// <summary>
         ///     The item's unique identifier.
         /// </summary>
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
 
         /// <summary>
         ///     The column's unique identifier.
